GC Fix: reduce refcount when returning from a method
After the CALL instruction, the result is pushed onto the stack (so its
refcount increases), but it already has a refcount of 1 because inside
the method it was PUSHed once. Result: it's never garbage collected.

The solution is to manually decrease the object refcount by one when it
comes from a CALL.